REFACTOR, PEP 621, AND OTHER.
The name of the repository have been changed to the final `aiida-vibroscopy`,
which is more appealing in respect to the previous one.
The source directory has been moved to `src/`, which is
becoming the standard recently, although PyPA is not yet
officially recommending it as the standard.
Following PEP 621 it is now possible to fully define the build procedure
of your package in pyproject.toml. Since this PEP is now well
supported, and for example pip can use it, we migrate to it since it
provides a bunch of benefits.
Furthermore, the all the names referring to `dph0_susceptibility` have
been changed to `raman_susceptibility`, which is more adherent to the
literature and is more readable.
Added `isort` and `flynt` to the pre-commit.