`Docs`: update deps and general improvements (#29)
Some dependencies are updated to get the newer
version of Myst-Nb, which was supporting only
Sphinx < 6.0 and was generally problematic for
e.g. pydantic module.

A change of theme style is also applied, to stay
on the same page as of aiida-qe in order to not
confuse users with different layouts.