`PhononWorkChain`: add max concurrent running pw workchains Fixes #52
Add the `settings.max_concurrent_base_workchains` input to the `PhononWorkChain`
to run only up to a maximum number of PwBaseWorkChain at the same time. This is
particularly useful for low symmetry materials where many displacements are generated.
This avoids to overload both the AiiDA daemon and especially the HPC. Moreover, this
even allows for running on the local machine (where AiiDA is installed) this workflow.