`Properties`: reduce mesh for phonon mesh interpolation (#98)
The value of 1000 for mesh interpolation is extreme and unnecessary.
Phonopy defaults it to 100, which corresponds already to very large
cells of about 50'000 atoms. We therefore choose 300, which is a
very reasonable value and it is still feasible to run on common laptops.

In passing, we fix the dependencies to make the tests to pass, and we
constrain the spglib version.